pi_backend/pyproject.toml
[tool.poetry]
name = "gatekeeper"
version = "0.1.0"
description = ""
authors = ["Stephen Onnen <stephen.onnen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^1.1"
Flask-Restful = "^0.3.7"
"RPi.GPIO" = "^0.7.0"
Gunicorn = "^19.9"
Flask-SQLAlchemy = "^2.4"
psycopg2-binary = "^2.8"
marshmallow = "^3.2"
flask_marshmallow = "^0.10.1"
marshmallow-sqlalchemy = "^0.19.0"
rpi_ws281x = "^4.2"
adafruit-circuitpython-neopixel = "^3.3"

[tool.poetry.dev-dependencies]
isort = "^4.3"
pytest = "^3.0"
black = { version = "*", allows-prereleases = true }
coverage = "^4.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
